ACS-8278: rename package prediction_applier to hxi_extension
