--- prediction-applier-extension/src/main/java/org/alfresco/hxi_connector/prediction_applier/rest/api/util/NodesUtils.java
@@ -1,44 +1,0 @@
-/*
- * #%L
- * Alfresco HX Insight Connector
- * %%
- * Copyright (C) 2023 - 2024 Alfresco Software Limited
- * %%
- * This file is part of the Alfresco software.
- * If the software was purchased under a paid Alfresco license, the terms of
- * the paid license agreement will prevail.  Otherwise, the software is
- * provided under the following open source license terms:
- *
- * Alfresco is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Alfresco is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
- * #L%
- */
-
-package org.alfresco.hxi_connector.prediction_applier.rest.api.util;
-
-import static lombok.AccessLevel.PRIVATE;
-
-import lombok.NoArgsConstructor;
-
-import org.alfresco.rest.api.Nodes;
-import org.alfresco.service.cmr.repository.NodeRef;
-
-@NoArgsConstructor(access = PRIVATE)
-public class NodesUtils
-{
-    // To keep compatibility with 7.3.x ACS versions we cannot use validateOrLookupNode(id)
-    public static NodeRef validateOrLookupNode(Nodes nodes, String id)
-    {
-        return nodes.validateOrLookupNode(id, null);
-    }
-}
