+++ prediction-applier-extension/src/main/resources/alfresco/module/alfresco-hxinsight-connector-prediction-applier-extension/context/service-context.xml
@@ -20,14 +20,14 @@
        xsi:schemaLocation="http://www.springframework.org/schema/beans
           http://www.springframework.org/schema/beans/spring-beans-3.0.xsd">
 
-    <bean id="predictionService" class="org.alfresco.hxi_connector.prediction_applier.service.PredictionServiceImpl">
+    <bean id="predictionService" class="org.alfresco.hxi_connector.hxi_extension.service.PredictionServiceImpl">
         <constructor-arg name="nodeService" ref="NodeService"/>
         <constructor-arg name="namespaceService" ref="NamespaceService"/>
     </bean>
 
     <bean id="PredictionService" class="org.springframework.aop.framework.ProxyFactoryBean">
         <property name="proxyInterfaces">
-            <value>org.alfresco.hxi_connector.prediction_applier.service.PredictionService</value>
+            <value>org.alfresco.hxi_connector.hxi_extension.service.PredictionService</value>
         </property>
         <property name="target">
             <ref bean="predictionService"/>
@@ -55,21 +55,21 @@
 
     <bean id="PredictionService_security" class="org.alfresco.repo.security.permissions.impl.AlwaysProceedMethodInterceptor"/>
 
-    <bean class="org.alfresco.hxi_connector.prediction_applier.rest.api.NodeEntityResource">
+    <bean class="org.alfresco.hxi_connector.hxi_extension.rest.api.NodeEntityResource">
         <property name="nodeService" ref="NodeService" />
         <property name="nodes" ref="nodes" />
         <property name="predictionService" ref="PredictionService"/>
     </bean>
 
-    <bean class="org.alfresco.hxi_connector.prediction_applier.rest.api.PredictionChildrenRelation">
+    <bean class="org.alfresco.hxi_connector.hxi_extension.rest.api.PredictionChildrenRelation">
         <property name="nodes" ref="nodes"/>
         <property name="transactionService" ref="TransactionService"/>
         <property name="predictionService" ref="PredictionService"/>
     </bean>
 
-    <bean class="org.alfresco.hxi_connector.prediction_applier.rest.api.QuestionsEntityResource"/>
+    <bean class="org.alfresco.hxi_connector.hxi_extension.rest.api.QuestionsEntityResource"/>
 
-    <bean class="org.alfresco.hxi_connector.prediction_applier.rest.api.PredictionsEntityResource">
+    <bean class="org.alfresco.hxi_connector.hxi_extension.rest.api.PredictionsEntityResource">
         <property name="nodes" ref="nodes"/>
         <property name="predictionService" ref="PredictionService"/>
     </bean>
@@ -77,7 +77,7 @@
     <bean id="hxi.simpleMappingExceptionResolver" abstract="true" parent="simpleMappingExceptionResolverParent">
         <property name="exceptionMappings">
             <map merge="true">
-                <entry key="org.alfresco.hxi_connector.prediction_applier.rest.api.exception.PredictionStateChangedException" value="#{T(org.springframework.extensions.webscripts.Status).STATUS_CONFLICT}" />
+                <entry key="org.alfresco.hxi_connector.hxi_extension.rest.api.exception.PredictionStateChangedException" value="#{T(org.springframework.extensions.webscripts.Status).STATUS_CONFLICT}" />
             </map>
         </property>
     </bean>
@@ -121,7 +121,7 @@
         </property>
     </bean>
 
-    <bean id="hxInsightAuthClient" class="org.alfresco.hxi_connector.prediction_applier.client.HxInsightAuthClient">
+    <bean id="hxInsightAuthClient" class="org.alfresco.hxi_connector.hxi_extension.client.HxInsightAuthClient">
         <constructor-arg name="authProperties" ref="hxiAuthProperties"/>
     </bean>
 
@@ -129,5 +129,5 @@
         <constructor-arg name="authenticationClient" ref="hxInsightAuthClient"/>
     </bean>
 
-    <bean class="org.alfresco.hxi_connector.prediction_applier.rest.api.QuestionsEntityResource"/>
+    <bean class="org.alfresco.hxi_connector.hxi_extension.rest.api.QuestionsEntityResource"/>
 </beans>
