+++ prediction-applier/src/main/java/org/alfresco/hxi_connector/prediction_applier/PredictionApplierApplication.java
@@ -29,8 +29,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan
 @EnableConfigurationProperties
 @ConfigurationPropertiesScan
 @SuppressWarnings("PMD.UseUtilityClass")
